Create File handling and columnar transposition.cpp
This piece of code works with 2 files: user.txt and system.txt. If option 1 is chosen, the text contained in user.txt is copied in system.txt. If option 2 is chosen and a list of words to be censored is given, the text is copied from user.txt to system.txt with those censored. If option 3 is chosen, the text from system.txt is encrypted. Option 4 is to be continued.